Shakti-main/requirements.txt
Flask
scapy
pyyaml
stem
requests
netifaces
psutil
flask-cors
